requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==4.0.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
